renames landingpage component to home component
