+++ src/main/routes/AppRouter/index.tsx
@@ -2,7 +2,7 @@ import React, { Suspense } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Route, Switch, withRouter } from "react-router";
 import ErrorBoundary from "../../components/Pages/ErrorBoundary";
-import { LandingPage } from "../../components/Pages/LandingPage";
+import { LandingPage } from "../../components/Pages/Home";
 import { LoadingPage } from "../../components/Pages/LoadingPage";
 // import { NotFoundPage } from "../../components/Pages/NotFoundPage";
 import { setDropDownAction } from "../../store/modules/dropDown/actions";
